Gandi: correctly handle simultaneous challenges

+++ gandi-livedns/cleanup.py
@@ -17,6 +17,7 @@
 import pprint
 import requests
 import os
+import sys
 from config import *
 
 pp = pprint.PrettyPrinter(indent=4)
@@ -26,14 +27,14 @@ try:
     certbot_domain
 except NameError:
     print("CERTBOT_DOMAIN environment variable is missing, exiting")
-    exit(1)
+    sys.exit(1)
 
 certbot_validation = os.environ.get("CERTBOT_VALIDATION")
 try:
     certbot_validation
 except NameError:
     print("CERTBOT_VALIDATION environment variable is missing, exiting")
-    exit(1)
+    sys.exit(1)
 
 if livedns_sharing_id is None:
     sharing_param = ''
@@ -51,7 +52,7 @@ if (response.ok):
     domains = response.json()
 else:
     response.raise_for_status()
-    exit(1)
+    sys.exit(1)
 
 domain_index = next((index for (index, d) in enumerate(domains) if d["fqdn"] == certbot_domain), None)
 
@@ -59,30 +60,43 @@ if domain_index is None:
     # domain not found
     print('The requested domain {} was not found in this gandi account'
           .format(certbot_domain))
-    exit(1)
+    sys.exit(1)
 
 domain_records_href = domains[domain_index]["domain_records_href"]
 
-response = requests.get('{}/_acme-challenge{}'
+response = requests.get('{}/_acme-challenge/TXT{}'
                         .format(domain_records_href, sharing_param),
                         headers=headers)
 
-if (response.ok):
-    domains = response.json()
-    if len(domains) == 0:
-        print("Existing _acme-challenge record not found, exiting")
-        exit(1)
-    if domains[0]["rrset_type"] != "TXT":
-        print('Existing _acme-challenge record doesn\'t seems to be '
-              'an acme challenge, exiting')
-else:
-    print("Failed to look for existing _acme-challenge record")
-    response.raise_for_status()
-    exit(1)
+if response.status_code == 404:
+    print('No _acme-challenge TXT record found.')
+    print('Nothing to cleanup, exiting')
+    sys.exit(0)
 
-response = requests.delete('{}/_acme-challenge/TXT{}'
-                           .format(domain_records_href, sharing_param),
-                           headers=headers)
+if not response.ok:
+    print('Failed to look for existing _acme-challenge record')
+    response.raise_for_status()
+    sys.exit(1)
+
+existing_record = response.json()
+if '"{}"'.format(certbot_validation) not in existing_record['rrset_values']:
+    print('Validation token not in _acme-challenge TXT record found.')
+    print('Nothing to cleanup, exiting')
+    sys.exit(0)
+
+# pp.pprint(existing_record)
+if len(existing_record.get('rrset_values', [])) > 1:
+    newrecord = existing_record
+    newrecord['rrset_values'].remove('"{}"'.format(certbot_validation))
+    # print(newrecord)
+    response = requests.put('{}/_acme-challenge/TXT{}'
+                            .format(domain_records_href, sharing_param),
+                            headers=headers, json=newrecord)
+else:
+    # Need to completely remove record
+    response = requests.delete('{}/_acme-challenge/TXT{}'
+                               .format(domain_records_href, sharing_param),
+                               headers=headers)
 
 if (response.ok):
     print("all good, entry deleted")
@@ -91,4 +105,4 @@ else:
     print("something went wrong")
     pp.pprint(response.content)
     response.raise_for_status()
-    exit(1)
+    sys.exit(1)
